requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Eel==0.18.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
func_timeout==4.3.5
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
psutil==7.0.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
PyMsgBox==1.0.9
pynput==1.8.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
pytweening==1.2.0
pywin32==309
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==76.0.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
